SNR, P and L std, float storage
Store the SNR as a float, rather than a uint16_t, avoiding the scaling
and rounding at each access.

The comnav and tersus decoders were not even using the SNR_UNIT for
scaling, perhaps another encoding, but assume scaling is not need when
stored as a float.

Store the P and L std in the float format. Makes better use of the
input format, without lossy compression. Avoids having to do the
encoding and decoding to the compressed format at each use.

Implement the compressed encoding in RINEX input and output. Easier to
make changes in one location here, and avoids this lossy compression
on other paths.

This uses a little extra storage, but if this is an issue then it
might be better to abstract a compressed encoding into accessor
functions than to have this distributed to each use. The P and L std
might be made a compilation option.

ublox: decode_trkmeas is estimating Lstd so have it respect the
-RCVSTDS option.

unicore: was not initialising P and L std.