str2str readcmd: avoid constant sizes
MAXSTR was not the appropriate definition of the buffer size here as
it is the number of streams which is distinct from the command line
read chunk size. Use a read chuck size of 81 characters.

Pass in the target cmd size, rather than depending on MAXRCVCMD.

Exit as soon as the target is full, rather than filling with smaller
later chucks, and emit a warning on overflow.

Only accept the separator character @ at the start of a line and
flush to the end of this line.