readerp: support IGS UT1-TAI offsets
The UT1-TAI offset is included in the IGS format, and products using
this offset are being produced. Canonicalised these offsets to UT1-UTC
for consistent internal use.

Also detect the IGS format as having 'VERSION 2' in the first line,
and distinguish between header and data lines, and support
concatenated files that include the version and header lines.