sunmoonpos: add a cache for performance
and include the original faster versions as a compilation option.

The calcuation of the sun position had a significant impact on
performance. It is called repeatedly with the same time so a cache is
effective.

The sun and moon calculations have been split into two separate
functions and it is possible to compile in the prior faster version of
either separately.