convbin rtcm time: use the start or end time if supplied
This RTCM time is a rough time needed to determine the absolute time
from the incomplete RTCM encoded time. This can be supplied by the
'-tr' argument but this will often be start time. So if not supplied
by this argument then use the supplied start or end time before
falling back to using the file time because the file time can change
if copied or processed.