winapp rtknavi: remove dup labels, update tab order, fix some options
Remove some duplicate labels or unused labels.

Improve the tab orderings.

Reorganise the GUI layout descriptions to be closer to the order of
presentation, closer to the tab ordering. The development tools do not
maintain an order, but it is tedious to try to use the graphical
development tools to obtain consistent layout.

Improve the layout alignments.

Add file loading for the BLQ file.

Fix the loading and saving of the BLQ and proxy address config options.

Fix viewing of the DCB file. The method was missing and it was using
the station position file.

Add rtklib.h to a few header files.

Correct solution sstat and trace options. The state for these was
duplicated in other variables and not always being set in the solopt_t
object. Remove this duplication to resolve this.

rtkconv: default to RINEX 3.04, and enable all signal types

rtkpost: increase the number of input files from 6 to 10, matching the
qt app.

Update the frequency dialog to include the 6th index and BDS B2ab.

srvtblbrows: Expand the cast list width to better view long
URLs. Strip trailing carriage returns from URLs.

Remove some unused variables, to quieten compiler warnings.