rtksvr: implement automated antenna selection, pos in static mode
Implement the automated antenna type selection along with the antenna
delta from the input stream, from an RTCM or raw stream, the config
options ant{1,2}-anttype="*".

Have the config options ant{1,2}-postype RTCM and raw automatically
set the position, and independently of the antenna type, and now
including for the rover when in fixed mode.

Previously only an antenna delta would be picked up as an offset to a
base position, including a RTCM height, and only when the position was
also being automatically set and only for the base and not the rover.

If the antenna type is empty then the delta is still applied but the
GUI apps were disabling input of the delta in this case.

If the GUI apps have the antenna type check box off then the type is
considered empty but the delta can still be supplied and will be
applied.

This allows a delta from a marker to an antenna ARP to be supplied in
the GUI apps even when the antenna does not have an antenna type
supplied.

When the antenna type is enabled in the GUI apps and set to '*' (for
auto) then the delta is still applied until it is updated from the
RTCM or raw stream, so it is still enabled in this case in the GUI to
allow setting an initial delta.

Winapp: Enable the rcv and sat PCV check-boxes in relative modes as
these can be applied, as already done for the qt app.

There was no separate 'raw' antenna automated position mode, this
is included under the 'rtcm' option.