winapp rtknavi: enable solution processing options
The rtksvr can now output the processing options in the header. Sync
with the qt app which recently enabled this.

Also enable the solution format in the options tab, as even though it
is not used by rtksvr it is used to gate the other options enabled and
those options may be used. The solution format rather comes from the
output streams, specific to each output stream. The user it left to
sync these options. This also syncs with the qt app.