gui console output cleanup
A few functions use similar code to output stream ASCII or hex to a
console. The code was adding linefeeds to wrap lines and then
detecting these, and also leaving them which was noise in most
cases. This reworks the code to avoid these trailing linefeeds.

rtknavi_qt was initializing the stream console output checkbox to
'HEX' but the initial output was ASCII.