rtkplot: add sky and DOP plots using the solution stat input
The solution stat file has the satellites used in the solution and the
satellite azimuth and elevation which is a usable input for these
plots, and without loading the observation and navigation data. It
also gives the data used in the solution. These are added as
additional plots to the current observation data based sky and DOP
plot which are still useful when examining this data and planning
solutions.