Automate source code formatting, using clang-format
The Google code style is adopted. This is a relatively terse style, giving
compact code, which appears to be the most consistent with original coding
style which was very compact.

The column limit is set at 100 characters. Much of the original code was
written with a column limit of around 80 characters, but was very terse with
little use of white space and reformatted this with more white space while
keeping a column limit of 80 would lead to excessive line wrapping, so the
column limit is relaxed to 100 characters. Developers are still encouraged to
write new C code with a line limit of 80 characters where practical.

A few places are exempted, using 'clang-format off/on', such as formatted
tables and include files that can not be reordered. Developers are encouraged
to avoid hand formatted code styles and work with the automatted formatting.

There is a lot of hand formatted code that is lost with this automated
formatting. There can be a loss of compactness of the code and a loss of
parallel code not being so well presented. What benefit this might have had
for readability is balanced against the maintenance and productivity burden.

Some code is styled with long trailing comments, or trailing comments at the
end of long lines, and that does not fit the 100 column limit. The automated
formatting does not hoist these, and some of these code comments might be
better moved to a line on their own.

Patches against code before the reformat can be reworked by applying the
automated formatting to the code before and after the patch and creating a new
diff. This reformatting is limited to just the automated reformatting of the
code to assist this transition.